feat: Add inventory file for Proxmox Hypervisor Clusters, Docker Installs, and VMs
This commit adds a new inventory file `hosts.yaml` that includes the configuration for Proxmox Hypervisor Clusters, Docker Installs, and VMs. It also includes new files and tasks related to these configurations.

The purpose of this commit is to provide a centralized inventory file that can be used for provisioning and managing the infrastructure components. This will make it easier to deploy and maintain the various services and resources.

The changes include:
- Adding configuration for Proxmox Hypervisor Clusters, Docker Installs, and VMs in the `hosts.yaml` file.
- Adding a new file `file.txt` in the `roles/service/files` directory, which can be used to copy files to hosts.
- Adding a new file `main.yaml` in the `roles/service/handles` directory, which includes a task to call the `echo` command.
- Adding a new file `external.yaml` in the `roles/service/tasks` directory, which includes a simple external task.
- Adding a new file `main.yaml` in the `roles/service/tasks` directory, which includes tasks for checking the existence of a file and importing an external task.
- Adding a new file `template.conf.j2` in the `roles/service/templates` directory, which includes a template for generating a configuration file.
- Adding a new file `main.yaml` in the `roles/service/vars` directory, which includes a variable `var_template` with the value "bar".

These changes will improve the organization and maintainability of the infrastructure configuration.

Note: The commit message has been generated based on the provided code changes. Please review and modify it as needed.